more separation. moving lazy feature to database module
