+++ h5rdmtoolbox/convention/standard_names/h5interface.py
@@ -134,7 +134,7 @@ class HDF5StandardNameInterface:
     @classmethod
     def from_hdf(cls, hdf_filename, group='/', recursive: bool = False):
         """search withing a group. Note, that duplicate standard names are not considered"""
-        from ...wrapper.lazy import lazy
+        from ...database.lazy import lazy
         hdf_filename = pathlib.Path(hdf_filename)
         standard_datasets = {}
         with h5tbx.File(hdf_filename) as h5:
