+++ h5rdmtoolbox/wrapper/__init__.py
@@ -4,6 +4,5 @@ Contains wrapper classes
 """
 
 from . import core
-from . import lazy
 
-__all__ = ['core', 'lazy']
+__all__ = ['core']
